backend/post-service/requirements.txt
Django
djangorestframework
djangorestframework-simplejwt
mysqlclient
django-cors-headers
gunicorn
Pillow